+++ app/cmd/main.fer
@@ -69,6 +69,9 @@ let scientific : f64 = 1.23e4; // 1.23 * 10^4
 //     };
 // }
 
+
+type Int i32;
+
 let cxx : data::MyType = data::kk;
 
 //let someValue = laterDeclaredValue;
