+++ app/cmd/start.fer
@@ -1,0 +1,17 @@
+let a: i32 = 10;
+let b = 23.232;
+
+let sum: str = a + b;
+
+// let x, y, z : i32; // all three variables are of type i32
+
+// let name, age, height : str, i32, f64 = "Alice", 25, 5.5; // multiple variable declaration with types
+
+
+fn doSomething() {
+    // This function does something
+}
+
+fn add(a: i32, b: i32) -> i32 {
+    return a + b;
+}
