+++ compiler/internal/report/reporter.go
@@ -20,6 +20,7 @@ type COMPILATION_PHASE string
 const (
 	LEXING_PHASE    COMPILATION_PHASE = "Lexing"
 	PARSING_PHASE   COMPILATION_PHASE = "Parsing"
+	COLLECTOR_PHASE COMPILATION_PHASE = "Collector"
 	RESOLVER_PHASE  COMPILATION_PHASE = "Resolver"
 	TYPECHECK_PHASE COMPILATION_PHASE = "Type Checking"
 )
